[Google Drive] Fix unwarranted deletion of root folders from files table
Description
---
See issue [5100](https://github.com/dust-tt/dust/issues/5100) more
specifically [here](https://github.com/dust-tt/dust/issues/5100#issuecomment-2137933661)

+++ connectors/src/connectors/google_drive/temporal/activities.ts
@@ -604,7 +604,15 @@ export async function garbageCollector(
           selectedFolders,
           lastSeenTs
         );
-        if (isInFolder === false || driveFile.trashed) {
+
+        const isRoot = !!(await GoogleDriveFolders.findOne({
+          where: {
+            connectorId: connectorId,
+            folderId: driveFile.id,
+          },
+        }));
+
+        if ((isInFolder === false && isRoot === false) || driveFile.trashed) {
           await deleteOneFile(connectorId, driveFile);
         } else {
           await file.update({
